requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Security
Flask-WTF==1.2.1
WTForms==3.1.1

# PGP Functionality
python-gnupg==0.5.2

# Production Server
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0

# Development Dependencies (optional, for local development)
# Uncomment for development:
# Flask-DebugToolbar==0.14.1
# pytest==7.4.3
# pytest-flask==1.3.0